// src/common/generated/prisma/enums.ts
export const OrganizationRole = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER'
} as const

export type OrganizationRole = (typeof OrganizationRole)[keyof typeof OrganizationRole]


export const TeamRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  DEVELOPER: 'DEVELOPER',
  VIEWER: 'VIEWER'
} as const

export type TeamRole = (typeof TeamRole)[keyof typeof TeamRole]